+++ src/runtime/components/Checkbox.vue
@@ -49,7 +49,7 @@ export interface CheckboxProps extends Pick<CheckboxRootProps, 'disabled' | 'req
 }
 
 export type CheckboxEmits = {
-  change: [payload: Event]
+  change: [event: Event]
 }
 
 export interface CheckboxSlots {