+++ src/runtime/components/Chip.vue
@@ -35,7 +35,7 @@ export interface ChipProps {
 }
 
 export interface ChipEmits {
-  'update:show': [payload: boolean]
+  'update:show': [value: boolean]
 }
 
 export interface ChipSlots {