+++ src/runtime/components/RadioGroup.vue
@@ -68,7 +68,7 @@ export interface RadioGroupProps<T extends RadioGroupItem = RadioGroupItem> exte
 }
 
 export type RadioGroupEmits = RadioGroupRootEmits & {
-  change: [payload: Event]
+  change: [event: Event]
 }
 
 type NormalizeItem<T extends RadioGroupItem> = Exclude<T & { id: string }, RadioGroupValue>